requirements.txt
# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0

# 数据库
sqlalchemy>=2.0.0
alembic>=1.12.0
aiosqlite>=0.19.0

# 异步支持
asyncio-mqtt>=0.16.0
aioredis>=2.0.0
aiocache>=0.12.0

# 任务调度
apscheduler>=3.10.0

# 数据验证
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 工具
pyyaml>=6.0
jinja2>=3.1.0
pillow>=10.0.0
playwright>=1.40.0
httpx>=0.25.0
beautifulsoup4>=4.12.0

# AI/LLM
dashscope>=1.14.0  # 阿里云通义千问
openai>=1.0.0  # 备用

# 图像处理
opencv-python>=4.8.0
numpy>=1.24.0

# CLI工具
click>=8.1.0
tabulate>=0.9.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# NoneBot2 接入
nonebot2>=2.2.0
nonebot-adapter-onebot>=2.4.3